Rename NewGameScreenXXX classes to NewGameXXX and extract helper classes to own file
